+++ src/Api/Dfe.Complete.Api/Controllers/TasksDataController.cs
@@ -285,6 +285,25 @@ namespace Dfe.Complete.Api.Controllers
             return NoContent();
         }
         
+
+        /// <summary>
+        /// Update the church supplemental agreement task Data for conversion or transfer project.
+        /// </summary>
+        /// <param name="request">The update command.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [Authorize(Policy = "CanReadWriteUpdate")]
+        [HttpPatch]
+        [Route("TaskData/ChurchSupplementalAgreement")]
+        [SwaggerResponse(204, "Conversion or transfer's church supplemental agreement task updated successfully.")]
+        [SwaggerResponse(400, "Invalid request data.")]
+        [SwaggerResponse(404, "Project/User not found.")]
+        public async Task<IActionResult> UpdateChurchSupplementalAgreementTaskAsync(
+            [FromBody] UpdateChurchSupplementalAgreementTaskCommand request,
+            CancellationToken cancellationToken)
+        {
+            await sender.Send(request, cancellationToken);
+            return NoContent();
+        }
         /// <summary>
         /// Confirm conversion project's land questionnaire task updated successfully.
         /// </summary>