+++ CHANGELOG.md
@@ -39,6 +39,7 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 - Added 'Closing of transfer or conversion project'.
 - Add API end point for creating user
 - Add ability to create a user
+- Added `Church supplemental agreement` task for both conversion and transfer projects.
 
 ### Changed